Add files via upload
Email Address analyzer - 
This analyzer can recognize email addresses located in any text. The addresses can be of any format. The analyzers extracts information like the local name, the domain name(the top-level-domain(tld), the country domain(cd)), the country name etc. and gives the output in JSON format.
Telephone analyzer-
This analyzer recognizes any telephone numbers that are a part of the text. The information of the area, station, prefix, the type of telephone and the original text are given in the output
Links analyzer- 
Similar to the address analyzer, this analyzer also identifies any hyperlinks in any format that are present in the text and extracts information like the scheme, the top level domain, the subdomain(if present), and the page path in the link.